+++ ChangeLog.md
@@ -4,6 +4,7 @@ This changelog summarizes changes and fixes which are a part of each revision.  
 commit history.
 
 * 2017/06/?? - 4.6 - PENDING
+   * site_data.yml completely changed.  Please update /app/prismeData.
    * Modified the labels for komet changing from rails_komet_a to Term Editor A on app deployer and on the home page
    * Fixed an IE bug in App Deployer regarding the description textarea placeholder which was not disappearing on focus of the field
    * Added application context validation message to App Deployer notifying the user if a given instance of KOMET is already deployed on a given Tomcat instance